refactor: 관리자용 storage 위치 변경
